api/requirements.txt
flask-sqlalchemy==2.5.1
flask-cors==3.0.10
flask-socketio==5.3.2
pyserial==3.5